Cleanup build.gradle
