+++ app/build.gradle
@@ -3,6 +3,28 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-parcelize'
 
+def versions = [
+        kotlin  : '1.4.31',
+        androidx: [
+                core            : '1.3.2',
+                appcompat       : '1.2.0',
+                constraintlayout: '2.0.4',
+                navigation      : '2.3.4',
+                lifecycle       : '2.3.0'
+        ],
+        material: '1.3.0',
+        timber  : '4.7.1',
+        koin    : '1.0.2',
+
+        test    : [
+                junit: '4.13.2',
+                androidx: [
+                        junit: '1.1.2',
+                        espresso : '3.3.0'
+                ]
+        ]
+]
+
 android {
     compileSdkVersion 30
     buildToolsVersion "30.0.3"
@@ -11,8 +33,8 @@ android {
         applicationId "com.netchar.nicknamer"
         minSdkVersion 23
         targetSdkVersion 30
-        versionCode 1
-        versionName "1.0"
+        versionCode extractVersionCode(appVersionName)
+        versionName appVersionName
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -50,25 +72,20 @@ android {
 }
 
 dependencies {
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$versions.kotlin"
+    implementation "androidx.core:core-ktx:$versions.androidx.core"
+    implementation "androidx.appcompat:appcompat:$versions.androidx.appcompat"
+    implementation "com.google.android.material:material:$versions.material"
+    implementation "androidx.constraintlayout:constraintlayout:$versions.androidx.constraintlayout"
+    implementation "androidx.navigation:navigation-fragment-ktx:$versions.androidx.navigation"
+    implementation "androidx.navigation:navigation-ui-ktx:$versions.androidx.navigation"
+    implementation "androidx.lifecycle:lifecycle-common-java8:$versions.androidx.lifecycle"
+    implementation "com.jakewharton.timber:timber:$versions.timber"
+    implementation "org.koin:koin-android:$versions.koin"
+    implementation "org.koin:koin-android-viewmodel:$versions.koin"
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.4'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.3.4'
-    implementation "androidx.lifecycle:lifecycle-common-java8:2.3.0"
-    implementation "com.jakewharton.timber:timber:4.7.1"
-    // koin version
-    def koin_version = '1.0.2'
-
-    // Koin for Android
-    implementation "org.koin:koin-android:$koin_version"
-    // or Koin for view model features
-    implementation "org.koin:koin-android-viewmodel:$koin_version"
-
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    // Tests
+    testImplementation "junit:junit:$versions.test.junit"
+    androidTestImplementation "androidx.test.ext:junit:$versions.test.androidx.junit"
+    androidTestImplementation "androidx.test.espresso:espresso-core:$versions.test.androidx.espresso"
 }
