+++ app/src/main/java/com/netchar/nicknamer/domen/models/Nickname.kt
@@ -3,7 +3,7 @@ package com.netchar.nicknamer.domen.models
 
 inline class Nickname(val value: String) {
     companion object {
-        val EMPTY = Nickname("")
+        private val EMPTY = Nickname("")
 
         fun Nickname?.orEmpty() = this ?: EMPTY
     }
