+++ app/src/main/java/com/netchar/nicknamer/domen/service/NicknameGeneratorService.kt
@@ -1,6 +1,5 @@
 package com.netchar.nicknamer.domen.service
 
-import com.netchar.nicknamer.domen.models.Config
 import com.netchar.nicknamer.domen.models.Nickname
 
 interface NicknameGeneratorService {
@@ -14,4 +13,13 @@ interface NicknameGeneratorService {
     fun removeFromFavorites(nickname: Nickname)
 
     fun getFavoriteNicknames(): List<Nickname>
+
+    data class Config(
+        val nicknameLength: Int = 5,
+        val gender: Gender = Gender.MALE,
+        val alphabet: Alphabet = Alphabet.LATIN
+    ) {
+        enum class Gender(val value: String) { MALE("male"), FEMALE("female") }
+        enum class Alphabet(val value: String) { LATIN("latin"), CYRILLIC("cyrillic") }
+    }
 }
