+++ app/src/main/java/com/netchar/nicknamer/domen/service/NicknameGeneratorServiceImpl.kt
@@ -1,6 +1,5 @@
 package com.netchar.nicknamer.domen.service
 
-import com.netchar.nicknamer.domen.models.Config
 import com.netchar.nicknamer.domen.models.Nickname
 import com.netchar.nicknamer.domen.NicknameModelsDataSource
 import com.netchar.nicknamer.domen.models.Nickname.Companion.orEmpty
@@ -11,7 +10,7 @@ class NicknameGeneratorServiceImpl(
     companion object {
         private const val PEOPLE_PREFIX = "people"
     }
-    override suspend fun generateNickname(config: Config): Nickname {
+    override suspend fun generateNickname(config: NicknameGeneratorService.Config): Nickname {
         val key = "${PEOPLE_PREFIX}_${config.gender.value}_${config.alphabet.value}"
         val source = dataSource.getDataSource()
         return source[key]?.generateNickname(config.nicknameLength).orEmpty()
